requirements.txt
asgiref==3.11.1
attrs==26.1.0
Django==6.0.3
djangorestframework==3.17.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
PyJWT==2.12.1
sqlparse==0.5.5
pytokens==0.4.1
requests==2.33.1
python-dateutil==2.9.0.post0
django-q2==1.9.0
